+++ lib/dart_trackers.dart
@@ -2,13 +2,17 @@
 library;
 
 export 'src/assignment.dart';
+export 'src/botsort_tracker.dart';
 export 'src/bytetrack_tracker.dart';
+export 'src/cbiou_tracker.dart';
 export 'src/cmc.dart';
+export 'src/converters.dart';
 export 'src/detections.dart';
 export 'src/iou.dart';
 export 'src/kalman_filter.dart';
 export 'src/matrix.dart';
-export 'src/botsort_tracker.dart';
+export 'src/motion_model.dart';
 export 'src/ocsort_tracker.dart';
 export 'src/sort_tracker.dart';
+export 'src/timing.dart';
 export 'src/tracker.dart';